@@ -14,44 +14,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.2.12",
-    "@angular/cdk": "^14.2.7",
-    "@angular/common": "^14.2.12",
-    "@angular/compiler": "^14.2.12",
-    "@angular/core": "^14.2.12",
-    "@angular/flex-layout": "^14.0.0-beta.41",
-    "@angular/forms": "^14.2.12",
-    "@angular/material": "^14.2.7",
-    "@angular/platform-browser": "^14.2.12",
-    "@angular/platform-browser-dynamic": "^14.2.12",
-    "@angular/router": "^14.2.12",
+    "@angular/animations": "^15.2.0",
+    "@angular/cdk": "^15.2.0",
+    "@angular/common": "^15.2.0",
+    "@angular/compiler": "^15.2.0",
+    "@angular/core": "^15.2.0",
+    "@angular/flex-layout": "^15.0.0-beta.42",
+    "@angular/forms": "^15.2.0",
+    "@angular/material": "^15.2.0",
+    "@angular/platform-browser": "^15.2.0",
+    "@angular/platform-browser-dynamic": "^15.2.0",
+    "@angular/router": "^15.2.0",
     "@auth0/angular-jwt": "^5.1.2",
     "@date-io/core": "1.3.7",
     "@date-io/date-fns": "1.3.7",
     "@flowjs/flow.js": "^2.14.1",
     "@flowjs/ngx-flow": "~0.6.0",
-    "@fullcalendar/angular": "^4.4.5-beta",
-    "@fullcalendar/core": "^4.4.0",
-    "@fullcalendar/daygrid": "^4.4.0",
-    "@fullcalendar/interaction": "^4.4.0",
-    "@fullcalendar/list": "^4.4.0",
-    "@fullcalendar/moment": "^4.4.0",
-    "@fullcalendar/timegrid": "^4.4.0",
+    "@fullcalendar/angular": "^6.1.4",
+    "@fullcalendar/core": "^6.1.4",
+    "@fullcalendar/daygrid": "^6.1.4",
+    "@fullcalendar/interaction": "^6.1.4",
+    "@fullcalendar/list": "^6.1.4",
+    "@fullcalendar/moment": "^6.1.4",
+    "@fullcalendar/timegrid": "^6.1.4",
     "@geoman-io/leaflet-geoman-free": "^2.13.0",
     "@juggle/resize-observer": "^3.3.1",
-    "@mat-datetimepicker/core": "~10.1.1",
+    "@mat-datetimepicker/core": "~11.0.3",
     "@material-ui/core": "4.12.3",
     "@material-ui/icons": "4.11.2",
     "@material-ui/pickers": "3.3.10",
-    "@ngrx/effects": "^14.3.3",
-    "@ngrx/store": "^14.3.3",
-    "@ngrx/store-devtools": "^14.3.3",
+    "@ngrx/effects": "^15.3.0",
+    "@ngrx/store": "^15.3.0",
+    "@ngrx/store-devtools": "^15.3.0",
     "@ngx-translate/core": "^14.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
     "@tinymce/tinymce-angular": "^7.0.0",
     "ace-builds": "1.4.13",
     "ace-diff": "^3.0.3",
-    "angular-gridster2": "~14.1.4",
+    "angular-gridster2": "~15.0.3",
     "angular2-hotkeys": "^13.1.0",
     "canvas-gauges": "^2.1.7",
     "core-js": "^3.26.1",
@@ -172,7 +174,7 @@
     "protractor": "~7.0.0",
     "raw-loader": "^4.0.2",
     "ts-node": "^10.9.1",
-    "typescript": "~4.6.4",
+    "typescript": "~4.8.4",
     "webpack": "^5.75.0"
   },
   "resolutions": {
