@@ -36,6 +36,7 @@ import {
   ElementRef,
   Input,
   OnChanges,
+  OnDestroy,
   OnInit,
   SimpleChanges,
   ViewChild,
@@ -60,18 +61,7 @@ import {
 import { CollectionViewer, DataSource, SelectionModel } from '@angular/cdk/collections';
 import { BehaviorSubject, forkJoin, fromEvent, merge, Observable, of, ReplaySubject } from 'rxjs';
 import { emptyPageData, PageData } from '@shared/models/page/page-data';
-import {
-  catchError,
-  debounceTime,
-  distinctUntilChanged,
-  map,
-  publishReplay,
-  refCount,
-  share,
-  skip,
-  take,
-  tap
-} from 'rxjs/operators';
+import { catchError, debounceTime, distinctUntilChanged, map, share, skip, take, tap } from 'rxjs/operators';
 import { PageLink, PageQueryParam } from '@shared/models/page/page-link';
 import { SchedulerEventService } from '@core/http/scheduler-event.service';
 import { MatPaginator } from '@angular/material/paginator';
@@ -79,7 +69,7 @@ import { MatSort, SortDirection } from '@angular/material/sort';
 import { Direction, SortOrder, sortOrderFromString } from '@shared/models/page/sort-order';
 import { UtilsService } from '@core/services/utils.service';
 import { TranslateService } from '@ngx-translate/core';
-import { deepClone, isDefined, isEmptyStr, isNotEmptyStr, isNumber } from '@core/utils';
+import { deepClone, isDefined, isEmptyStr, isNotEmptyStr, isNumber, isString } from '@core/utils';
 import { MatDialog } from '@angular/material/dialog';
 import {
   SchedulerEventDialogComponent,
@@ -94,7 +84,7 @@ import dayGridPlugin from '@fullcalendar/daygrid';
 import listPlugin from '@fullcalendar/list';
 import timeGridPlugin from '@fullcalendar/timegrid';
 import momentPlugin, { toMoment } from '@fullcalendar/moment';
-import interactionPlugin from '@fullcalendar/interaction';
+import interactionPlugin, { DateClickArg } from '@fullcalendar/interaction';
 import { FullCalendarComponent } from '@fullcalendar/angular';
 import {
   schedulerCalendarView,
@@ -102,12 +92,8 @@ import {
   schedulerCalendarViewValueMap,
   SchedulerEventsWidgetSettings
 } from '@home/components/scheduler/scheduler-events.models';
-import { Calendar, DateClickApi } from '@fullcalendar/core/Calendar';
-import { asRoughMs, Duration, EventInput, rangeContainsMarker } from '@fullcalendar/core';
-import { EventSourceError, EventSourceInput } from '@fullcalendar/core/structs/event-source';
-import * as _moment from 'moment';
+import { Calendar, CalendarOptions, Duration, EventClickArg, EventDropArg, EventInput, } from '@fullcalendar/core';
 import { MatMenuTrigger } from '@angular/material/menu';
-import { EventHandlerArg } from '@fullcalendar/core/types/input-types';
 import { getUserZone } from '@shared/models/time/time.models';
 import { ActivatedRoute, QueryParamsHandling, Router } from '@angular/router';
 import {
@@ -117,7 +103,8 @@ import {
 import { EntityType } from '@shared/models/entity-type.models';
 import { ResizeObserver } from '@juggle/resize-observer';
 import { hidePageSizePixelValue } from '@shared/models/constants';
-import * as moment_ from 'moment';
+import { asRoughMs, rangeContainsMarker } from '@fullcalendar/core/internal';
+import * as _moment from 'moment';
 
 @Component({
   selector: 'tb-scheduler-events',
@@ -125,7 +112,7 @@ import * as moment_ from 'moment';
   styleUrls: ['./scheduler-events.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
-export class SchedulerEventsComponent extends PageComponent implements OnInit, AfterViewInit, OnChanges {
+export class SchedulerEventsComponent extends PageComponent implements OnInit, AfterViewInit, OnChanges, OnDestroy {
 
   @ViewChild('schedulerEventWidgetContainer', {static: true}) schedulerEventWidgetContainerRef: ElementRef;
   @ViewChild('searchInput') searchInputField: ElementRef;
@@ -161,12 +148,6 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
 
   mode = 'list';
 
-  displayCreatedTime = true;
-  displayType = true;
-  displayCustomer = true;
-
-  schedulerEventConfigTypes: {[eventType: string]: SchedulerEventConfigType};
-
   displayPagination = true;
   pageSizeOptions;
   defaultPageSize = 10;
@@ -177,33 +158,27 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
 
   displayedColumns: string[];
   pageLink: PageLink;
-
   textSearchMode = false;
 
   assignEnabled = false;
 
   dataSource: SchedulerEventsDatasource;
 
-  calendarPlugins = [interactionPlugin, momentPlugin, dayGridPlugin, listPlugin, timeGridPlugin];
-
   currentCalendarView = schedulerCalendarView.month;
 
-  currentCalendarViewValue = schedulerCalendarViewValueMap.get(this.currentCalendarView);
-
-  schedulerCalendarViews = Object.keys(schedulerCalendarView);
+  schedulerCalendarViews = Object.keys(schedulerCalendarView) as schedulerCalendarView[];
   schedulerCalendarViewTranslations = schedulerCalendarViewTranslationMap;
 
-  eventSources: EventSourceInput[] = [this.eventSourceFunction.bind(this)];
-
-  calendarApi: Calendar;
-
-  schedulerEventMenuPosition = { x: '0px', y: '0px' };
-
+  schedulerEventMenuPosition = {x: '0px', y: '0px'};
   schedulerContextMenuEvent: MouseEvent;
 
-  private schedulerEvents: Array<SchedulerEventWithCustomerInfo> = [];
+  calendarOptions: CalendarOptions;
 
+  private calendarApi: Calendar;
+  private schedulerEvents: Array<SchedulerEventWithCustomerInfo> = [];
+  private currentCalendarViewValue = schedulerCalendarViewValueMap.get(this.currentCalendarView);
   private widgetResize$: ResizeObserver;
+  private schedulerEventConfigTypes: {[eventType: string]: SchedulerEventConfigType};
 
   constructor(protected store: Store<AppState>,
               private utils: UtilsService,
@@ -249,7 +224,7 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
         this.pageLink.textSearch = decodeURI(routerQueryParams.textSearch);
       }
       this.schedulerEventConfigTypes = deepClone(defaultSchedulerEventConfigTypes);
-      this.dataSource = new SchedulerEventsDatasource(this.schedulerEventService, this.schedulerEventConfigTypes, this.route);
+      this.dataSource = new SchedulerEventsDatasource(this.schedulerEventService, this.schedulerEventConfigTypes);
       if (this.edgeId) {
         const isEdgeWriteAllowed: boolean = this.userPermissionsService.hasGenericPermission(Resource.EDGE, Operation.WRITE);
         this.assignEnabled = isEdgeWriteAllowed;
@@ -277,6 +252,28 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
       });
       this.widgetResize$.observe(this.schedulerEventWidgetContainerRef.nativeElement);
     }
+    this.calendarOptions = {
+      plugins: [
+        interactionPlugin,
+        momentPlugin,
+        dayGridPlugin,
+        listPlugin,
+        timeGridPlugin
+      ],
+      height: '100%',
+      initialView: this.currentCalendarViewValue,
+      allDaySlot: false,
+      editable: this.editEnabled,
+      headerToolbar: false,
+      selectable: false,
+      eventDisplay: 'block',
+      eventDurationEditable: false,
+      events: this.eventSourceFunction.bind(this),
+      eventClick: this.onEventClick.bind(this),
+      dateClick: this.onDayClick.bind(this),
+      eventDrop: this.onEventDrop.bind(this),
+      eventDidMount: this.onEventDidMount.bind(this)
+    };
   }
 
   ngOnDestroy(): void {
@@ -379,7 +376,7 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
         }
       }
     ];
-    this.dataSource = new SchedulerEventsDatasource(this.schedulerEventService, this.schedulerEventConfigTypes, this.route);
+    this.dataSource = new SchedulerEventsDatasource(this.schedulerEventService, this.schedulerEventConfigTypes);
     this.dataSource.selection.changed.subscribe(() => {
       const hideTitlePanel = !this.dataSource.selection.isEmpty() || this.textSearchMode;
       if (this.ctx.hideTitlePanel !== hideTitlePanel) {
@@ -413,7 +410,8 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
               this.updateData();
             } else {
               const queryParams: PageQueryParam = {
-                textSearch: encodeURI(this.pageLink.textSearch) || null,
+                textSearch: isString(this.pageLink.textSearch) && this.pageLink.textSearch !== ''
+                  ? encodeURI(this.pageLink.textSearch) : null,
                 page: null
               };
               this.updatedRouterQueryParams(queryParams);
@@ -441,12 +439,10 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
       if (this.displayPagination) {
         if (this.displayPagination) {
           paginatorSubscription$ = this.paginator.page.asObservable().pipe(
-            map((data) => {
-              return {
-                page: data.pageIndex === 0 ? null : data.pageIndex,
-                pageSize: data.pageSize === this.defaultPageSize ? null : data.pageSize
-              };
-            })
+            map((data) => ({
+              page: data.pageIndex === 0 ? null : data.pageIndex,
+              pageSize: data.pageSize === this.defaultPageSize ? null : data.pageSize
+            }))
           );
         }
       }
@@ -473,7 +469,6 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
             this.textSearchMode = true;
             this.pageLink.textSearch = decodeURI(params.textSearch);
           } else {
-            this.textSearchMode = false;
             this.pageLink.textSearch = null;
           }
           this.updateData();
@@ -590,7 +585,7 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
   }
 
   addSchedulerEvent($event: Event) {
-      this.openSchedulerEventDialog($event);
+    this.openSchedulerEventDialog($event);
   }
 
   assignToEdgeSchedulerEvent($event: Event) {
@@ -603,8 +598,8 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
     }
     this.schedulerEventService.getSchedulerEvent(schedulerEventWithCustomerInfo.id.id)
       .subscribe((schedulerEvent) => {
-      this.openSchedulerEventDialog($event, schedulerEvent);
-    });
+        this.openSchedulerEventDialog($event, schedulerEvent);
+      });
   }
 
   viewSchedulerEvent($event, schedulerEventWithCustomerInfo: SchedulerEventWithCustomerInfo) {
@@ -617,7 +612,7 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
       });
   }
 
-  openSchedulerEventDialog($event: Event, schedulerEvent?: SchedulerEvent, readonly = false) {
+  private openSchedulerEventDialog($event: Event, schedulerEvent?: SchedulerEvent, readonly = false) {
     if ($event) {
       $event.stopPropagation();
     }
@@ -657,7 +652,7 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
     );
   }
 
-  openAssignSchedulerEventToEdgeDialog($event: Event) {
+  private openAssignSchedulerEventToEdgeDialog($event: Event) {
     if ($event) {
       $event.stopPropagation();
     }
@@ -680,7 +675,7 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
 
   triggerResize() {
     setTimeout(() => {
-      this.calendarComponent.getApi().updateSize();
+      this.calendarApi.updateSize();
     }, 0);
   }
 
@@ -703,9 +698,8 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
 
   isCalendarToday(): boolean {
     if (this.calendarApi) {
-      const now = this.calendarApi.getNow();
       const view = this.calendarApi.view;
-      return rangeContainsMarker(view.props.dateProfile.currentRange, now);
+      return rangeContainsMarker({start: view.activeStart, end: view.activeEnd}, Date.now());
     } else {
       return false;
     }
@@ -719,14 +713,14 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
     this.calendarApi.next();
   }
 
-  onEventClick(arg: EventHandlerArg<'eventClick'>) {
+  private onEventClick(arg: EventClickArg) {
     const schedulerEvent = this.schedulerEvents.find(event => event.id.id === arg.event.id);
     if (schedulerEvent) {
       this.openSchedulerEventContextMenu(arg.jsEvent, schedulerEvent);
     }
   }
 
-  openSchedulerEventContextMenu($event: MouseEvent, schedulerEvent: SchedulerEventWithCustomerInfo) {
+  private openSchedulerEventContextMenu($event: MouseEvent, schedulerEvent: SchedulerEventWithCustomerInfo) {
     $event.preventDefault();
     $event.stopPropagation();
     const $element = $(this.calendarContainer.nativeElement);
@@ -736,7 +730,7 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
     this.schedulerContextMenuEvent = $event;
     this.schedulerEventMenuPosition.x = x + 'px';
     this.schedulerEventMenuPosition.y = y + 'px';
-    this.schedulerEventMenuTrigger.menuData = { schedulerEvent };
+    this.schedulerEventMenuTrigger.menuData = {schedulerEvent};
     this.schedulerEventMenuTrigger.openMenu();
   }
 
@@ -744,7 +738,7 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
     this.schedulerEventMenuTrigger.closeMenu();
   }
 
-  onDayClick(event: DateClickApi) {
+  private onDayClick(event: DateClickArg) {
     if (this.addEnabled) {
       const calendarDate = new Date(event.date.getTime() + event.date.getTimezoneOffset() * 60 * 1000);
       const date = toMoment(calendarDate, this.calendarApi);
@@ -763,43 +757,16 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
     }
   }
 
-  onEventDrop(arg: EventHandlerArg<'eventDrop'>) {
+  private onEventDrop(arg: EventDropArg) {
     const schedulerEvent = this.schedulerEvents.find(event => event.id.id === arg.event.id);
     if (schedulerEvent) {
       this.moveEvent(schedulerEvent, arg.delta, arg.revert);
     }
   }
 
-  private moveEvent(event: SchedulerEventWithCustomerInfo, delta: Duration, revertFunc: () => void) {
-    this.schedulerEventService.getSchedulerEvent(event.id.id).subscribe(
-      (schedulerEvent) => {
-        schedulerEvent.schedule.startTime += asRoughMs(delta);
-        this.schedulerEventService.saveSchedulerEvent(schedulerEvent).subscribe(
-          () => {
-            this.reloadSchedulerEvents();
-          },
-          () => {
-            revertFunc();
-          }
-        );
-      },
-      () => {
-        revertFunc();
-      }
-    );
-  }
-
-  eventRender(arg: EventHandlerArg<'eventRender'>) {
-    const props: {name: string, type: string, info: string, htmlTitle: string} = arg.event.extendedProps;
-    if (props.htmlTitle) {
-      if (arg.el.getElementsByClassName('fc-title').length > 0) {
-        arg.el.getElementsByClassName('fc-title')[0].innerHTML = props.htmlTitle;
-      }
-      if (arg.el.getElementsByClassName('fc-list-item-title').length > 0) {
-        arg.el.getElementsByClassName('fc-list-item-title')[0].innerHTML = props.htmlTitle;
-      }
-    }
-    const element = $(arg.el);
+  private onEventDidMount(event: any) {
+    const props: { name: string; type: string; info: string; repeatInterval: string } = event.event.extendedProps;
+    const element = $(event.el);
     import('tooltipster').then(
       () => {
         element.tooltipster(
@@ -822,26 +789,45 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
         );
         const tooltip = element.tooltipster('instance');
         tooltip.content($(
-          '<div class="tb-scheduler-tooltip-title">' + props.name + '</div>' +
-          '<div class="tb-scheduler-tooltip-content"><b>' + this.translate.instant('scheduler.event-type') + ':</b> ' + props.type + '</div>' +
-          '<div class="tb-scheduler-tooltip-content">' + props.info + '</div>'
+          `<div class="tb-scheduler-tooltip-title">${props.name}</div>` +
+          `<div class="tb-scheduler-tooltip-content"><b>${this.translate.instant('scheduler.event-type')}:&nbsp;</b>${props.type}</div>` +
+          `<div class="tb-scheduler-tooltip-content">${props.info}</div>`
         ));
       }
     );
   }
 
-  updateCalendarEvents(schedulerEvents: Array<SchedulerEventWithCustomerInfo>) {
+  private moveEvent(event: SchedulerEventWithCustomerInfo, delta: Duration, revertFunc: () => void) {
+    this.schedulerEventService.getSchedulerEvent(event.id.id).subscribe({
+      next: (schedulerEvent) => {
+        schedulerEvent.schedule.startTime += asRoughMs(delta);
+        this.schedulerEventService.saveSchedulerEvent(schedulerEvent).subscribe({
+          next: () => {
+            this.reloadSchedulerEvents();
+          },
+          error: () => {
+            revertFunc();
+          }
+        });
+      },
+      error: () => {
+        revertFunc();
+      }
+    });
+  }
+
+  private updateCalendarEvents(schedulerEvents: Array<SchedulerEventWithCustomerInfo>) {
     this.schedulerEvents = schedulerEvents;
     if (this.calendarApi) {
       this.calendarApi.refetchEvents();
     }
   }
 
-  eventSourceFunction(arg: {
+  private eventSourceFunction(arg: {
     start: Date;
     end: Date;
     timeZone: string;
-  },                  successCallback: (events: EventInput[]) => void, failureCallback: (error: EventSourceError) => void) {
+  }, successCallback: (events: EventInput[]) => void, failureCallback: (error: Error) => void) {
     const events: EventInput[] = [];
     if (this.schedulerEvents && this.schedulerEvents.length) {
       const start = toMoment(arg.start, this.calendarApi);
@@ -850,17 +836,13 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
       const rangeStart = start.local();
       const rangeEnd = end.local();
       this.schedulerEvents.forEach((event) => {
-        const startOffset = userZone.utcOffset(event.schedule.startTime) * 60 * 1000;
-        const eventStart = _moment(event.schedule.startTime - startOffset);
+        const eventStart = _moment(event.schedule.startTime);
         let calendarEvent: EventInput;
         if (rangeEnd.isSameOrAfter(eventStart)) {
           if (event.schedule.repeat) {
-            const endOffset = userZone.utcOffset(event.schedule.repeat.endsOn) * 60 * 1000;
-            const repeatEndsOn = _moment(event.schedule.repeat.endsOn - endOffset);
+            const repeatEndsOn = _moment(event.schedule.repeat.endsOn);
             if (event.schedule.repeat.type === SchedulerRepeatType.TIMER) {
-              calendarEvent = this.toCalendarEvent(event,
-                eventStart,
-                repeatEndsOn);
+              calendarEvent = this.toCalendarEvent(event, eventStart, repeatEndsOn);
               events.push(calendarEvent);
             } else {
               let currentTime: _moment.Moment;
@@ -874,27 +856,24 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
                   case SchedulerRepeatType.EVERY_N_DAYS:
                   case SchedulerRepeatType.EVERY_N_WEEKS:
                     const eventStartOffsetDuration = _moment.duration(rangeStart.diff(eventStart));
-                    let offsetUnits: moment_.unitOfTime.Base;
+                    let offsetUnits: _moment.unitOfTime.Base;
                     if (event.schedule.repeat.type === SchedulerRepeatType.EVERY_N_DAYS) {
                       offsetUnits = 'days';
-                      eventStartOffsetUnits =
-                        Math.ceil(eventStartOffsetDuration.as(offsetUnits));
+                      eventStartOffsetUnits = Math.ceil(eventStartOffsetDuration.as(offsetUnits));
                       const remainder = eventStartOffsetUnits % event.schedule.repeat.days;
                       if (remainder) {
                         eventStartOffsetUnits += event.schedule.repeat.days - remainder;
                       }
                     } else if (event.schedule.repeat.type === SchedulerRepeatType.EVERY_N_WEEKS) {
                       offsetUnits = 'weeks';
-                      eventStartOffsetUnits =
-                        Math.ceil(eventStartOffsetDuration.as(offsetUnits));
+                      eventStartOffsetUnits = Math.ceil(eventStartOffsetDuration.as(offsetUnits));
                       const remainder = eventStartOffsetUnits % event.schedule.repeat.weeks;
                       if (remainder) {
                         eventStartOffsetUnits += event.schedule.repeat.weeks - remainder;
                       }
                     } else {
                       offsetUnits = schedulerRepeatTypeToUnitMap.get(event.schedule.repeat.type);
-                      eventStartOffsetUnits =
-                        Math.ceil(eventStartOffsetDuration.as(offsetUnits));
+                      eventStartOffsetUnits = Math.ceil(eventStartOffsetDuration.as(offsetUnits));
                     }
                     currentTime = eventStart.clone().add(eventStartOffsetUnits, offsetUnits);
                     break;
@@ -928,13 +907,11 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
                     break;
                   case SchedulerRepeatType.EVERY_N_DAYS:
                     eventStartOffsetUnits += event.schedule.repeat.days;
-                    currentTime = eventStart.clone()
-                      .add(eventStartOffsetUnits, 'days');
+                    currentTime = eventStart.clone().add(eventStartOffsetUnits, 'days');
                     break;
                   case SchedulerRepeatType.EVERY_N_WEEKS:
                     eventStartOffsetUnits += event.schedule.repeat.weeks;
-                    currentTime = eventStart.clone()
-                      .add(eventStartOffsetUnits, 'weeks');
+                    currentTime = eventStart.clone().add(eventStartOffsetUnits, 'weeks');
                     break;
                   default:
                     currentTime.add(1, 'days');
@@ -955,13 +932,12 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
 
   private toCalendarEvent(event: SchedulerEventWithCustomerInfo, start: _moment.Moment, end?: _moment.Moment): EventInput {
     const title = `${event.name} - ${event.typeName}`;
-    let htmlTitle = null;
+    let repeatInterval;
     if (event.schedule.repeat && event.schedule.repeat.type === SchedulerRepeatType.TIMER) {
-      const repeatInterval = this.translate.instant(schedulerTimeUnitRepeatTranslationMap.get(event.schedule.repeat.timeUnit),
+      repeatInterval = this.translate.instant(schedulerTimeUnitRepeatTranslationMap.get(event.schedule.repeat.timeUnit),
         {count: event.schedule.repeat.repeatInterval});
-      htmlTitle = ` <b>${repeatInterval}</b> ${title}`;
     }
-    const calendarEvent: EventInput = {
+    return {
       id: event.id.id,
       title,
       name: event.name,
@@ -969,9 +945,8 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
       info: this.eventInfo(event),
       start: start.toDate(),
       end: end ? end.toDate() : null,
-      htmlTitle
+      repeatInterval
     };
-    return calendarEvent;
   }
 
   private eventInfo(event: SchedulerEventWithCustomerInfo): string {
@@ -996,9 +971,9 @@ export class SchedulerEventsComponent extends PageComponent implements OnInit, A
       } else if (event.schedule.repeat.type === SchedulerRepeatType.YEARLY) {
         info += this.translate.instant('scheduler.yearly') + ', ';
       } else if (event.schedule.repeat.type === SchedulerRepeatType.TIMER) {
-          const repeatInterval = this.translate.instant(schedulerTimeUnitRepeatTranslationMap.get(event.schedule.repeat.timeUnit),
-            {count: event.schedule.repeat.repeatInterval});
-          info += repeatInterval + ', ';
+        const repeatInterval = this.translate.instant(schedulerTimeUnitRepeatTranslationMap.get(event.schedule.repeat.timeUnit),
+          {count: event.schedule.repeat.repeatInterval});
+        info += repeatInterval + ', ';
       } else {
         info += this.translate.instant('scheduler.weekly') + ' ' + this.translate.instant('scheduler.on') + ' ';
         event.schedule.repeat.repeatOn.forEach((day) => {
@@ -1080,8 +1055,7 @@ class SchedulerEventsDatasource implements DataSource<SchedulerEventWithCustomer
   public edgeId: string;
 
   constructor(private schedulerEventService: SchedulerEventService,
-              private schedulerEventConfigTypes: {[eventType: string]: SchedulerEventConfigType},
-              private route: ActivatedRoute) {
+              private schedulerEventConfigTypes: { [eventType: string]: SchedulerEventConfigType }) {
   }
 
   connect(collectionViewer: CollectionViewer):
@@ -1154,8 +1128,12 @@ class SchedulerEventsDatasource implements DataSource<SchedulerEventWithCustomer
           });
           return schedulerEvents;
         }),
-        publishReplay(1),
-        refCount()
+        share({
+          connector: () => new ReplaySubject(1),
+          resetOnError: false,
+          resetOnComplete: false,
+          resetOnRefCountZero: false
+        })
       );
     }
     return this.allEntities;
