@@ -32,11 +32,18 @@
 import {
   ChangeDetectionStrategy,
   ChangeDetectorRef,
-  Component, ElementRef,
-  EventEmitter, HostBinding, Inject,
-  Input, OnDestroy,
+  Component,
+  ElementRef,
+  EventEmitter,
+  HostBinding,
+  Inject,
+  Input,
+  OnDestroy,
   OnInit,
-  Output, Renderer2, ViewChild
+  Output,
+  Renderer2,
+  ViewChild,
+  ViewEncapsulation
 } from '@angular/core';
 import { PageComponent } from '@shared/components/page.component';
 import { DashboardWidget, DashboardWidgets } from '@home/models/dashboard-component.models';
@@ -44,7 +51,7 @@ import { Store } from '@ngrx/store';
 import { AppState } from '@core/core.state';
 import { SafeStyle } from '@angular/platform-browser';
 import { WidgetExportType, widgetExportTypeTranslationMap } from '@shared/models/widget.models';
-import { guid, hashCode, isNotEmptyStr } from '@core/utils';
+import { guid, isNotEmptyStr } from '@core/utils';
 import cssjs from '@core/css/css';
 import { DOCUMENT } from '@angular/common';
 import { GridsterItemComponent } from 'angular-gridster2';
@@ -68,6 +75,7 @@ export class WidgetComponentAction {
   selector: 'tb-widget-container',
   templateUrl: './widget-container.component.html',
   styleUrls: ['./widget-container.component.scss'],
+  encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class WidgetContainerComponent extends PageComponent implements OnInit, OnDestroy {
